♻️ G9DW6M task: split preflight report helpers
